servisler tamam repositoryler dashboard ve giris css de

+++ WebApplication1/Repositories/GenericRepository.cs
@@ -1,11 +1,13 @@
-﻿using Microsoft.Extensions.Options;
-using MongoDB.Driver;
-using WebApplication1.Models;
+﻿using MongoDB.Driver;
+using MongoDB.Bson;
+using Microsoft.Extensions.Options;
 using WebApplication1.Repositories.Interfaces;
+using WebApplication1.Models;
+using System.Collections.Generic;
 
 namespace WebApplication1.Repositories
 {
-    public class GenericRepository<T> : IRepository<T>
+    public class GenericRepository<T> : IRepository<T> where T : class
     {
         private readonly IMongoCollection<T> _collection;
 
@@ -20,12 +22,15 @@ namespace WebApplication1.Repositories
 
         public List<T> GetAll()
         {
-            return _collection.Find(Builders<T>.Filter.Empty).ToList();
+            return _collection.Find(_ => true).ToList();
         }
 
         public T GetById(string id)
         {
-            var filter = Builders<T>.Filter.Eq("_id", id);
+            if (!ObjectId.TryParse(id, out ObjectId objectId))
+                return null;
+
+            var filter = Builders<T>.Filter.Eq("_id", objectId);
             return _collection.Find(filter).FirstOrDefault();
         }
 
@@ -36,13 +41,19 @@ namespace WebApplication1.Repositories
 
         public void Update(string id, T entity)
         {
-            var filter = Builders<T>.Filter.Eq("_id", id);
+            if (!ObjectId.TryParse(id, out ObjectId objectId))
+                return;
+
+            var filter = Builders<T>.Filter.Eq("_id", objectId);
             _collection.ReplaceOne(filter, entity);
         }
 
         public void Delete(string id)
         {
-            var filter = Builders<T>.Filter.Eq("_id", id);
+            if (!ObjectId.TryParse(id, out ObjectId objectId))
+                return;
+
+            var filter = Builders<T>.Filter.Eq("_id", objectId);
             _collection.DeleteOne(filter);
         }
     }
